Josh submission v1.1
Server and multiple clients implemented.

chatServer.java must be launched first to launch the server.

SecurityGUI.java and SecurityGUI2.java can be launched in any order. Each file generates a client and connects to the server. 

To send a message, type it in the 'inputText' textfield and click 'send'. It should then be passed to the server where it is parsed. The server then sends it back out through an output stream where both clients should read the message and display it in the 'Output' textarea.

Temperamental as fuck tho.